# docs/openapi/os-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Oficina Mecânica — OS API",
    "description": "Serviço OS da Fase 4: cadastros, autenticação de funcionários, abertura, status e histórico da ordem de serviço.",
    "version": "v1"
  },
  "paths": {
    "/os/ordens-servico/acompanhar/{numero}": {
      "get": {
        "tags": [
          "AcompanhamentoOS"
        ],
        "summary": "Acompanhamento da OS pelo número, pelo próprio cliente (JWT emitido pela Lambda de CPF).",
        "description": "Sem IDs internos: só número, status, datas e histórico. OS de outro cliente responde 404.",
        "parameters": [
          {
            "name": "numero",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AcompanhamentoOSResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcompanhamentoOSResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcompanhamentoOSResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/os/filiais": {
      "get": {
        "tags": [
          "Filiais"
        ],
        "summary": "Lista as filiais ativas, para informar o filialId na abertura da OS.",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilialResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilialResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilialResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/os/ordens-servico": {
      "get": {
        "tags": [
          "OrdensServico"
        ],
        "summary": "Lista as OS em andamento, da mais urgente (EmExecucao) para a menos (EmDiagnostico).",
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListarOrdensServicoItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListarOrdensServicoItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ListarOrdensServicoItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "OrdensServico"
        ],
        "summary": "Abre uma OS para cliente, veículo e filial. A OS nasce em EmDiagnostico.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbrirOrdemServicoRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AbrirOrdemServicoRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AbrirOrdemServicoRequest"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "OS criada; corpo traz o ID e o número consultáveis.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AbrirOrdemServicoResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbrirOrdemServicoResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbrirOrdemServicoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Campo obrigatório ausente; nenhuma OS é criada.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Veículo ou filial inexistente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Veículo não pertence ao cliente ou filial inativa.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AbrirOrdemServicoRequest": {
        "type": "object",
        "properties": {
          "clienteId": {
            "type": "string",
            "format": "uuid"
          },
          "veiculoId": {
            "type": "string",
            "format": "uuid"
          },
          "filialId": {
            "type": "string",
            "format": "uuid"
          },
          "observacoes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AbrirOrdemServicoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "numero": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AcompanhamentoOSResponse": {
        "type": "object",
        "properties": {
          "numero": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/StatusOrdemServico"
          },
          "dataAbertura": {
            "type": "string",
            "format": "date-time"
          },
          "dataFechamento": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "historico": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoricoStatusOSItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FilialResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HistoricoStatusOSItem": {
        "type": "object",
        "properties": {
          "statusAnterior": {
            "$ref": "#/components/schemas/StatusOrdemServico"
          },
          "statusNovo": {
            "$ref": "#/components/schemas/StatusOrdemServico"
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ListarOrdensServicoItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "numero": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/StatusOrdemServico"
          },
          "filialId": {
            "type": "string",
            "format": "uuid"
          },
          "dataAbertura": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "StatusOrdemServico": {
        "enum": [
          "EmDiagnostico",
          "AguardandoAprovacao",
          "AguardandoPagamento",
          "EmExecucao",
          "Finalizada",
          "Entregue",
          "Cancelada"
        ],
        "type": "string"
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT de funcionário (login) ou de cliente (Lambda de CPF).",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyInterna": {
        "type": "apiKey",
        "description": "Chave de API interna, usada só por /os/interno/*.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    },
    {
      "ApiKeyInterna": []
    }
  ]
}